Exception handling - delete

+++ src/main/java/com/wendellyv/course/services/UserService.java
@@ -3,8 +3,13 @@ package com.wendellyv.course.services;
 import java.util.List;
 import java.util.Optional;
 
+import com.wendellyv.course.services.exceptions.DatabaseException;
 import com.wendellyv.course.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
+import org.springframework.dao.DataAccessResourceFailureException;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.wendellyv.course.entities.User;
@@ -30,7 +35,15 @@ public class UserService {
 	}
 
 	public void delete(Long id){
-		repository.deleteById(id);
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException(id);
+		}
+
+		try {
+			repository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
 
 	public User update(Long id, User obj){
